preserve fractional spear knockback
Vanilla stab attacks add a built-in 0.4F extra knockback bonus before enchantment knockback. The local knockback feature only accepted an int amount, and the spear code rounded enchantment knockback plus 0.4F. With no Knockback enchantment that rounded to zero, so plain piercing and kinetic spear stabs lost the vanilla shove entirely.

The fix widens attack knockback to a double and passes the spear bonus through without rounding. Spear stabs now count knockback-only and dismount-only hits as affected like Player#stabAttack does, and they apply durability loss plus attack exhaustion per affected target instead of tying those side effects only to successful damage.

Piercing spear attacks also now run the Lunge post-piercing effect outside the hit-only block. Vanilla PiercingWeapon calls postPiercingAttack after the trace regardless of whether anything was hit, so a Lunge-enchanted spear can consume the Lunge durability/exhaustion and apply its impulse on a miss while a plain spear miss still does not lose durability.

+++ src/main/java/io/github/togar2/pvp/feature/knockback/FairKnockbackFeature.java
@@ -54,7 +54,7 @@ public class FairKnockbackFeature extends VanillaKnockbackFeature {
 
 	@Override
 	protected boolean applyKnockback(LivingEntity target, Entity attacker, @Nullable Entity source,
-	                                 EntityKnockbackEvent.KnockbackType type, int extraKnockback,
+	                                 EntityKnockbackEvent.KnockbackType type, double extraKnockback,
 	                                 double dx, double dz, boolean legacy) {
 		if (!(target instanceof Player player) || player.getLatency() < PING_OFFSET)
 			return super.applyKnockback(target, attacker, source, type, extraKnockback, dx, dz, legacy);
